CHROMIUM: cpuquiet: Move cpuquiet stats code to sysfs.c
It makes more sense for the logic that keeps track of and exposes stats
to sysfs to be in sysfs.c.

BUG=chrome-os-partner:44148
TEST=Boot smaug

Reviewed-on: https://chromium-review.googlesource.com/294183
