+++ drivers/cpuquiet/Makefile
@@ -1,4 +1,4 @@
 GCOV_PROFILE := y
 
-obj-$(CONFIG_CPU_QUIET) += cpuquiet.o driver.o governor.o governors/
+obj-$(CONFIG_CPU_QUIET) += cpuquiet.o governor.o governors/
 obj-$(CONFIG_CPU_QUIET_STATS) += sysfs.o attribute.o
